requirements.txt
# Core Web Framework
flask>=2.3.0,<3.0
gunicorn>=21.2.0,<22.0
werkzeug>=2.3.0,<3.0

# Data Processing
pandas>=2.2.0,<3.0
numpy>=1.26.0,<2.0

# Visualization & AI
plotly>=6.1.1,<7.0
kaleido>=0.2.1,<0.3.0
google-generativeai>=0.6.0,<0.7.0

# Serialization & Utilities
orjson>=3.9.10,<4.0
networkx>=3.2.1,<4.0

# Excel & Document Handling
openpyxl>=3.1.2,<4.0
python-docx>=1.0.0,<2.0
reportlab>=4.0.8,<5.0

# PDF & OCR & table extraction (heavy)
camelot-py[cv]>=0.11.0,<0.12.0
pdfplumber>=0.10.0,<0.11.0
pytesseract>=0.3.10,<0.4.0
opencv-python-headless>=4.8.0,<5.0

# Extras that may be useful for parsing / fonts / images
Pillow>=10.0.0,<11.0
python-magic>=0.4.27,<0.5.0

# Optional but recommended for speed
pyarrow>=12.0.1,<13.0

# Keep pip, setuptools modern
pip>=23.0
setuptools>=68.0
wheel>=0.40.0

python-dotenv>=1.0.0,<2.0

